Prevent user-select in html
Dragging occasionally has a problem with pulling an unknown item (which
appears to be the page icon). This prevents mouse click events from
firing properly, and leaving dragged elements stuck halfway.

The setting of user-select to none might possibly stop this, but there
it cannot really be tested. It has worked well so far so it is assumed
to be fixing the cause for now.