Introduce calculation of available moves
Add in the 'ChessEngine', 'ChessEngineAdapter', 'Piece' and its
subclasses, as well as helper classes of 'Point' and 'Grid'.

The ChessEngine makes use of a Grid to keep track of the board state,
and handles checking of which moves are valid in that board state.
Piece subclasses are there to specify what moves can be made for the
different chess pieces.

As ChessEngine makes use of Point to specify the position of each piece,
a ChessEngineAdapter is used to convert Point based responses into
square numbers as used by the UserInterface.

As a result of computing valid moves, they are then represented on the
board using move-markers that are inserted into the DOM.

Added JSDoc comments to specify the functions of the newly added
classes and functions, and for UserInterface. To add further
documentation in the future.