Streamline chess board initialization
Refactored reading of plan (now renamed as layout) into ChessEngine
and UserInterface.

The layout of pieces populate elements in the DOM directly, as well as
input pieces into the HTML DOM. Having this separate
initialization might cause errors if they are populated incorrectly and
are out of sync. Instead, have UserInterface read the board state
directly from the initialized ChessEngine.

As a result, any errors in the DOM would come directly from ChessEngine
and not from the layout.

Renamed plan to be layout. Refactored event listeners to be added after
everything else. Moved layout from ChessEngine constructor to an init
function, and set precomputeMoves to only be called privately.