Add handling of moves for chess pieces
Updates the 'move' function of the ChessEngine to actually change the
board state.

After the grid has been changed, the ui can proceed to update the DOM
to mirror the grid.

There is a difference between dropping a piece and clicking on the
destination square. The 'setPiecePosition' function has been modified
to handle the animation of pieces in each situation, using the 'slide'
CSS class.

Reformatted CSS files, and fixed a bug with ui.selected being 0 does
not request a move.