Simplify move object implementation
Replaced move objects in ChessEngine with a Move class. Removed unused
isCapturing property of Move.

Moves were previously defined using objects manually created. The
moveType property of a move can play the role of the isCapturing
property, so can remove it entirely.

As a result, moves are defined in one place, and with one less property
needed.

ChessEngineAdapter however continues using a new object, as the
properties of Move still need to be modified.