Fix bug with pawn movement
Remembered to change the moved property of pieces to true after moving
them.

The pawn was supposed to use this property to check if it was moving
for the first time, and let it move twice in that case. Updated pawn to
make use of this.

Now pawns should correctly move 1 square after having moved the first
time.