Implement castling move
Adds a new 'MoveType' of 'CASTLE', as well as modified existing methods
to support castling.

Castling requires the movement of two pieces instead of one. This
results in the need to pass more information in a move, so that
'ChessEngine' and 'UserInterface' can move the appropriate pieces. This
unfortunately results in the addition of extra statements to both
classes and the adapter class to support moving two pieces. A new type
of move 'CastleMove' extends 'Move' to include this additional
information.

As a result, castling can be done thought at the expense of increasing
the number of move types that need to be handled. This will be so for
one additional move type for en-passant which will be added soon. While
less elegant adding two new move types to check for regular chess
should be less costly than rewriting the existing engine to be more
extensible.

The order of setting a piece's 'moved' property is wrong, changed to be
before precomputing new moves.