Fix castling bug
Add additional condition for skipping validation in 'validatePosition'.

Moves are validated by checking if enemy pieces can capture king.
However squares travelled by king are not registered as 'capture' moves
since king is not present on those squares. Remove the skip condition
when castling so that the normal 'move' can be checked against king
positions.

Previous implementation seemed to work when enemy piece was a 'capture'
move by coincidence as it was not capturing the king, but the rook in
its end position.