Implement pawn promotion
Create a promotion window for player to select when moving pawn to last
rank.

Game needed to wait for additional player selection when promoting
(choosing which piece to promote to). Delayed the call to 'requestMove'
and called it later on making selection in promotion window.

Fixed bug with Grid where moving piece to the same point effectively
deleted it.